backend/requirements.txt
arrow==1.4.0
asgiref==3.11.1
betterpath==0.2.2
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==6.0.2
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-tailwind==4.4.2
djangorestframework==3.16.1
idna==3.11
Jinja2==3.1.6
l==0.11.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
vcversioner==2.16.0.0
zope.interface==8.2
gunicorn==24.1.0
psycopg2-binary>=2.9.10
dj-database-url==2.3.0
whiteNoise==6.9.0
